requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.4
django-ckeditor==6.4.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-oauth-toolkit==1.7.1
django-simpleui==2022.4.9
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
jwcrypto==1.0
lxml==4.8.0
MarkupPy==1.14
MarkupSafe==2.1.1
moviepy==1.0.3
munch==2.5.0
mysqlclient==2.1.0
Naked==0.1.31
numpy==1.22.4
oauthlib==3.2.0
odfpy==1.4.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.0
proglog==0.1.10
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.14.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
weixin-python==0.5.7
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
